**Overhaul CI/CD pipeline and modernize codebase**
- Rewrite GitHub Actions workflow to include comprehensive testing matrix (OS/Python versions), Ruff linting/formatting, and PyPI deployment automation
- Update documentation configuration and project metadata
- Refactor core modules to use pathlib, improve type hints, and enhance error handling
- Migrate from versioneer to manual version management
- Standardize development tooling with Ruff replacing flake8/isort
- Improve test coverage and validation logic

This commit modernizes the project infrastructure, improves maintainability, and prepares for reliable deployment workflows. The changes span CI configuration, code quality improvements, version management, and testing enhancements.

+++ ffsubsync/golden_section_search.py
@@ -1,6 +1,7 @@
 """Python program for golden section search (straight-up copied from Wikipedia).
-   This implementation reuses function evaluations, saving 1/2 of the evaluations per
-   iteration, and returns a bounding interval."""
+This implementation reuses function evaluations, saving 1/2 of the evaluations per
+iteration, and returns a bounding interval."""
+
 import logging
 import math
 
@@ -35,7 +36,7 @@ def gss(f, a, b, tol=1e-4):
         return a, b
 
     # Required steps to achieve tolerance
-    n = int(math.ceil(math.log(tol / h) / math.log(invphi)))
+    n = math.ceil(math.log(tol / h) / math.log(invphi))
     logger.info(
         "About to perform %d iterations of golden section search to find the best framerate",
         n,
